requirements.txt
numpy
pyelftools
telnetlib3
progressbar2